requirements.txt
aiogram==3.3.0
requests
python-dotenv
asyncpg
fastapi
uvicorn
asyncpg
python-dotenv
requests
aiogram
apscheduler